+++ spec/benchable_spec.rb
@@ -62,8 +62,8 @@ RSpec.describe Benchable do
     before do
       allow(bench_mock).to receive(:config)
       allow(bench_mock).to receive(:compare!)
-      allow(bench_mock).to receive(:report).with("Sum").and_yield
-      allow(bench_mock).to receive(:report).with("Multiplication").and_yield
+      allow(bench_mock).to receive(:report).with("sum").and_yield
+      allow(bench_mock).to receive(:report).with("multiplication").and_yield
       allow(Benchmark).to receive(:ips).and_yield(bench_mock)
     end
 
@@ -76,8 +76,8 @@ RSpec.describe Benchable do
     it "builds and runs a benchmark", :aggregate_failures do
       run_benchmark
 
-      expect(bench_mock).to have_received(:report).with("Sum")
-      expect(bench_mock).to have_received(:report).with("Multiplication")
+      expect(bench_mock).to have_received(:report).with("sum")
+      expect(bench_mock).to have_received(:report).with("multiplication")
       expect(bench_mock).to have_received(:compare!)
     end